test(org-browser): move reload persistence test to desktop only - W-23237574
Developer: Reload Window does a full page navigation in VS Code web,
re-fetching extension bundles from the dev web server — a source of
flakiness unrelated to what the test verifies (workspaceState
persistence across extension deactivate/reactivate). Desktop reloads
reliably, so move the test there and drop the web-only timeout
workarounds it required.
