+++ packages/salesforcedx-vscode-org-browser/test/playwright/pages/orgBrowserPage.ts
@@ -41,16 +41,11 @@ export class OrgBrowserPage {
     await waitForWorkspaceReady(this.page, 15_000);
   }
 
-  /**
-   * Open the Org Browser by clicking its activity bar item
-   * @param activityBarTimeout Timeout for the activity bar item to appear. Web mode needs a longer
-   * timeout right after a reload, since extension activation (and the context keys the view's
-   * `viewsWelcome` entries depend on) is slower there than on desktop.
-   */
-  public async openOrgBrowser(activityBarTimeout = 15_000): Promise<void> {
+  /** Open the Org Browser by clicking its activity bar item */
+  public async openOrgBrowser(): Promise<void> {
     await this.waitForProject();
     await expect(this.activityBarItem, 'Activity bar item for Org Browser should be visible').toBeVisible({
-      timeout: activityBarTimeout
+      timeout: 15_000
     });
 
     // Trigger navigation to Org Browser and wait for the types response
