+++ packages/salesforcedx-vscode-org-browser/test/playwright/playwright.config.web.ts
@@ -4,6 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import { defineConfig } from '@playwright/test';
 import { createWebConfig } from '@salesforce/playwright-vscode-ext';
 
-export default createWebConfig({ testDir: './specs' });
+export default defineConfig({ ...createWebConfig({ testDir: './specs' }), testIgnore: ['**/*.desktop.spec.ts'] });
